+++ ui/components/QuoteSection.less
