+++ ui/components/QuoteSection.tsx
