+++ ui/store/quote/reducer.ts
